remove unecessary code
