+++ Makefile
@@ -29,7 +29,6 @@ SRC			=	temp_functions.c \
 				is_operator.c \
 				get_pipedes.c \
 				get_commands.c \
-				clear_pipe_data.c \
 				exec_commands.c \
 				close_pipes.c \
